# OpenAPI.yaml
openapi: 3.0.0
info:
  title: Store
  version: '1.0'
  description: A simple store
  contact:
    name: SecuritEase Dev
    url: 'https://www.securitease.com'
    email: internal@securitease.com
  termsOfService: 'https://www.securitease.com'
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.htm'
servers:
  - url: 'http://localhost:8080'
    description: dev
paths:
  /api/v1/order:
    get:
      summary: Get Orders (Paginated)
      tags:
        - Order
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Zero-based page number (negative values default to 0)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Page size (capped at 100, values ≤ 0 default to 20)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOrderResponse'
      operationId: getOrders
      description: Get orders with pagination. Returns a bounded page of orders with their associated customer and products data.
    post:
      summary: Create Order
      tags:
        - Order
      responses:
        '201':
          description: Created
        '400':
          description: Invalid input — validation failed
        '404':
          description: Customer not found
      operationId: createOrder
      description: Create a new order
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
  /api/v1/order/{id}:
    get:
      summary: Get Order by ID
      tags:
        - Order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
      operationId: getOrderById
      description: Get a specific order by its ID
  /api/v1/customer:
    get:
      summary: Get Customers
      tags:
        - Customer
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Filter customers by name substring (case-insensitive)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
      operationId: getCustomers
      description: Get all customers, optionally filtered by name
    post:
      summary: Create Customer
      tags:
        - Customer
      responses:
        '201':
          description: Created
        '400':
          description: Invalid input — validation failed
      operationId: createCustomer
      description: Create a new customer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
  /api/v1/product:
    get:
      summary: Get Products
      tags:
        - Product
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
      operationId: getProducts
      description: Get all products with associated order IDs
    post:
      summary: Create Product
      tags:
        - Product
      responses:
        '201':
          description: Created
        '400':
          description: Invalid input — validation failed
      operationId: createProduct
      description: Create a new product
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
  /api/v1/product/{id}:
    get:
      summary: Get Product by ID
      tags:
        - Product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
      operationId: getProductById
      description: Get a specific product by its ID with associated order IDs
components:
  schemas:
    Order:
      title: Order
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        customer:
          type: object
          properties:
            id:
              type: integer
              format: int64
            name:
              type: string
        products:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
              description:
                type: string
    Customer:
      title: Customer
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        orders:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
              description:
                type: string
    CreateOrderRequest:
      title: CreateOrderRequest
      type: object
      required:
        - description
        - customerId
      properties:
        description:
          type: string
        customerId:
          type: integer
          format: int64
    CreateCustomerRequest:
      title: CreateCustomerRequest
      type: object
      required:
        - name
      properties:
        name:
          type: string
    Product:
      title: Product
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        orderIds:
          type: array
          items:
            type: integer
            format: int64
    CreateProductRequest:
      title: CreateProductRequest
      type: object
      required:
        - description
      properties:
        description:
          type: string
    PaginatedOrderResponse:
      title: PaginatedOrderResponse
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        pagination:
          $ref: '#/components/schemas/PaginationMetadata'
    PaginationMetadata:
      title: PaginationMetadata
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
          description: Total number of orders across all pages
        totalPages:
          type: integer
          description: Total number of pages
        currentPage:
          type: integer
          description: Current zero-based page number
        pageSize:
          type: integer
          description: Effective page size (capped at 100)
tags:
  - name: Customer
    description: Customer management operations
  - name: Order
    description: Order management operations
  - name: Product
    description: Product operations